+++ addons/miscFixes/patchRHSAFRF/CfgVehicles.hpp
@@ -1,0 +1,9 @@
+class CfgVehicles {
+    class B_AssaultPack_Base;
+    class rhs_medic_bag: B_AssaultPack_Base {};
+    class rhs_medic_bag_XL: rhs_medic_bag {
+        author = "RHS & AChesheireCat";
+        displayName = "Medic Bag (Large)";
+        maximumLoad = 240;
+    };
+};
